TileButton: fix accidentally visible label when labelVisible=false
If labelVisible=false, the $buttonLabel should never be shown, even
when a label is set. This corresponds to the behavior of most form
fields, but a normal Button (which is the parent class of TileButton)
behaves differently. In order not to break existing applications, we
only adjust this for TileButtons, and only if they exist inside a
dashboard tile.

Jasmine tests were added to test all combinations of label,
labelVisible, iconId and menus. Tests were also added to ButtonSpec
to assert the current (slightly unexpected) behavior.

393740