SequenceBox: mark as focused when a field without label is focused
FormFields are marked as "focused" when they have the focus. This causes
their form field label to be highlighted. If a field without a label
is placed in a SequenceBox with a label, the sequence box label was not
highlighted. But because the label visually appears to belong to the
field while it technically belongs to the sequence box, this is expected
by the user.

This change listens for the "focused" property of inner fields and
propagates it to the sequence box, as long as the inner field does not
have a visible label of its own. As soon as a visible label is
encountered, the sequence box is no longer marked.

Special case:
Labels that only contain a single punctuation character in are treated
as if their label was empty. This is useful for "from - to" sequence
boxes that only use a "-" sign as label for the second field.

The default behavior can be overwritten by adding marker classes:
- consider-inner-focus (sequence box)
- never-consider-inner-focus (sequence box)
- consider-for-outer-focus (inner field)
- never-consider-for-outer-focus (inner field)

390352